+++ Boinst.NAntExtensions.TeamCity/Tasks/SetBuildNumberTask.cs
@@ -1,0 +1,24 @@
+﻿namespace Boinst.NAntExtensions.TeamCity
+{
+    using NAnt.Core;
+    using NAnt.Core.Attributes;
+
+    [TaskName("teamcity-setbuildnumber")]
+    public class SetBuildNumberTask : Task
+    {
+        /// <summary>
+        /// The build number
+        /// </summary>
+        [TaskAttribute("value")]
+        public string BuildNumber
+        {
+            get;
+            set;
+        }
+
+        protected override void ExecuteTask()
+        {
+            this.Log(Level.Info, "##teamcity[buildNumber '{0}']", this.BuildNumber);
+        }
+    }
+}